[Rails 4.2] Support ActiveRecord 4.2
