+++ lib/activerecord_enum.rb
@@ -1,70 +1,10 @@
 require 'active_record'
 require 'active_record/base'
-require 'active_record/connection_adapters/abstract/schema_definitions.rb'
 
-require 'connection_adapters/sqlite3' if defined?( SQLite3 )
-require 'connection_adapters/mysql2' if defined?( Mysql2 )
+ACTIVE_RECORD_VERSION = ActiveRecord::VERSION
 
-module ActiveRecord
-  module ConnectionAdapters
-    class Column
-      def initialize_with_enum name, default, sql_type=nil, *args
-        initialize_without_enum name, default, sql_type, *args
-        @type = simplified_type_with_enum sql_type
-        @limit = extract_limit_with_enum sql_type
-        @default = extract_default_with_enum default
-      end
-      alias_method :initialize_without_enum, :initialize
-      alias_method :initialize, :initialize_with_enum
-
-      def simplified_type_with_enum field_type
-        if field_type =~ /enum|set/i
-          $&.to_sym
-        else
-          simplified_type field_type
-        end
-      end
-
-      def extract_limit_with_enum field_type
-        if field_type =~ /(?:enum|set)\(([^)]+)\)/i
-          $1.scan( /'([^']*)'/ ).flatten
-        else
-          extract_limit field_type
-        end
-      end
-
-      def extract_default_with_enum default
-        if type == :set
-          default.split "," if default.present?
-        else
-          extract_default default
-        end
-      end
-
-      def set?
-        type == :set
-      end
-
-      def enum?
-        type == :enum
-      end
-    end
-  end
-end
-module ActiveRecord
-  module ConnectionAdapters
-    class TableDefinition
-      def enum *args
-        options = args.extract_options!
-        column_names = args
-        column_names.each { |name| column(name, :enum, options) }
-      end
-      def set *args
-        options = args.extract_options!
-        options[:default] = options[:default].join "," if options[:default].present?
-        column_names = args
-        column_names.each { |name| column(name, :set, options) }
-      end
-    end
-  end
+if ACTIVE_RECORD_VERSION::MAJOR < 4 || (ACTIVE_RECORD_VERSION::MAJOR == 4 && ACTIVE_RECORD_VERSION::MINOR <= 1)
+  require 'activerecord_enum/activerecord_enum_pre42.rb'
+else
+  require 'activerecord_enum/activerecord_enum_post42.rb'
 end
